добавил подтверждение удаления, перенес формы, изменил css

+++ stocks/views.py
@@ -4,8 +4,10 @@ from django.shortcuts import render, get_object_or_404
 from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.db import connection
+from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from .models import Stocks
-from .forms import SearchForm, StockInfoForm
+from .forms import SearchForm, DealForm
 from .scripts.get_stocks import get_stocks_dict, get_stock_dict
 
 # Create your views here.
@@ -43,15 +45,8 @@ def stocks_main(request):
 
 def stock_detail(request, secid: str):
     stock = get_object_or_404(Stocks, secid=secid)
-    if request.method == 'POST':
-        form = StockInfoForm(request.POST, instance=stock)
-        if form.is_valid():
-            form.save()
-            redirect_url = reverse(f'stock_detail', args=[secid])
-            return HttpResponseRedirect(redirect_url)
-    else:
-        form = StockInfoForm(instance=stock)
-    values_list = list(stock.__dict__.items())[1:]
+    form = DealForm
+    values_list = list(stock.__dict__.items())[2:]
     context = {'values_list': values_list,
                'stock': stock,
                'form': form}
@@ -61,10 +56,9 @@ def stock_detail(request, secid: str):
 def stocks_search(request):
     form = SearchForm(request.GET)
     if form.is_valid():
-        if form.cleaned_data['field'] == 'Secname':
-            stocks = Stocks.objects.filter(secname__icontains=form.cleaned_data['input'])
-        elif form.cleaned_data['field'] == 'Secid':
-            stocks = Stocks.objects.filter(secid__icontains=form.cleaned_data['input'])
+            stocks = Stocks.objects.filter(Q(secid__icontains=form.cleaned_data['input']) |
+                                           Q(shortname__icontains=form.cleaned_data['input']) |
+                                           Q(secname__icontains=form.cleaned_data['input']))
     else:
         redirect_url = reverse('main_stocks')
         return HttpResponseRedirect(redirect_url)
@@ -77,32 +71,6 @@ def stocks_search(request):
 
 
 def stocks_update(request):
-    # stocks_fields_securities = get_stocks_dict(dict(StockInfoSecurities().__dict__.items()))
-    # stocks_fields_marketdata = get_stocks_dict(dict(StockInfoMarketdata().__dict__.items()))
-    # # обновить данные для таблицы StockInfoSecurities
-    # unchangeable_stocks = [item.secid for item in StockInfoSecurities.objects.all() if item.unchangeable is True]
-    # for item in stocks_fields_securities:
-    #     if item['secid'] not in unchangeable_stocks:
-    #         try:
-    #             StockInfoSecurities.objects.filter(secid=item['secid']).update(**item)
-    #         except:
-    #             StockInfoSecurities.objects.create(**item)
-    #     else:
-    #         del item['shortname']
-    #         del item['secname']
-    #         StockInfoSecurities.objects.filter(secid=item['secid']).update(**item)
-    # # обновить данные для таблицы StockInfoMarketdata
-    # for item in stocks_fields_marketdata:
-    #     try:
-    #         StockInfoMarketdata.objects.filter(secid=item['secid']).update(**item)
-    #     except:
-    #         StockInfoMarketdata.objects.create(**item)
-    # # связать таблицы у новых акций
-    # new_stocks = [item for item in StockInfoSecurities.objects.all() if not item.marketdata]
-    # if new_stocks:
-    #     for item in new_stocks:
-    #         item.marketdata = StockInfoMarketdata.objects.get(secid=f'{item.secid}')
-    #         item.save()
     auto_update()
     redirect_url = reverse('stocks_main')
     return HttpResponseRedirect(redirect_url)
@@ -115,16 +83,20 @@ def stock_update(request, secid: str):
     return HttpResponseRedirect(redirect_url)
 
 
+@login_required
 def stocks_delete(request):
-    with (connection.cursor() as cursor):
-        cursor.execute("delete from stocks_stocks")
-    redirect_url = reverse('stocks_main')
-    return HttpResponseRedirect(redirect_url)
+    if request.user.is_superuser:
+        with (connection.cursor() as cursor):
+            cursor.execute("delete from stocks_stocks")
+        redirect_url = reverse('stocks_main')
+        return HttpResponseRedirect(redirect_url)
 
 
+@login_required
 def stocks_download(request):
-    stocks_fields_securities = get_stocks_dict(dict(Stocks().__dict__.items()))
-    for item in stocks_fields_securities:
-        Stocks.objects.create(**item)
-    redirect_url = reverse('stocks_main')
-    return HttpResponseRedirect(redirect_url)
+    if request.user.is_superuser:
+        stocks_fields_securities = get_stocks_dict(dict(Stocks().__dict__.items()))
+        for item in stocks_fields_securities:
+            Stocks.objects.create(**item)
+        redirect_url = reverse('stocks_main')
+        return HttpResponseRedirect(redirect_url)
